[jmdb] Adding basic insert

+++ src/main/java/jmdb/spikes/mongodb/InsertAHugeNumberOfDocuments.java
@@ -3,9 +3,14 @@ package jmdb.spikes.mongodb;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DB;
 import com.mongodb.DBCollection;
+import com.mongodb.DBObject;
 import com.mongodb.MongoException;
+import com.mongodb.WriteResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static jmdb.spikes.mongodb.DataGenerator.nextContact;
 import static jmdb.spikes.mongodb.MongoDatabase.connectToDB;
@@ -15,8 +20,10 @@ public class InsertAHugeNumberOfDocuments implements Command {
 
     private static final Logger log = LoggerFactory.getLogger(InsertAHugeNumberOfDocuments.class);
 
-    private static int LOGGING_INTERVAL = 10;
-    private static String COLLECTION_NAME = "Addresses";
+    private static final int BATCH_SIZE = 5000;
+    private static final int TOTAL_TO_INSERT = 10000000;
+
+    private static final String COLLECTION_NAME = "AddressList";
 
     public void execute() {
         DB db = connectToDB();
@@ -24,21 +31,27 @@ public class InsertAHugeNumberOfDocuments implements Command {
         try {
 
             DBCollection collection = db.getCollection(COLLECTION_NAME);
-            log.info("dropping [%s] ...", COLLECTION_NAME);
-            collection.drop();
-            log.info("complete.");
-
-            collection = db.getCollection(COLLECTION_NAME);
+            log.info("Count = {}", collection.count());
+            if (collection.count() > 0) {
+                log.info("dropping [{}] ...", COLLECTION_NAME);
+                collection.drop();
+                log.info("complete.");
+            }
 
             log.info("starting insert ...");
-            for (int i = 0; i < 1000; ++i) {
-                BasicDBObject document = new BasicDBObject();
-                document.putAll(nextContact());
-                if (i % LOGGING_INTERVAL == 0) {
-                    log.info("Count = i");
+
+            int batches = TOTAL_TO_INSERT / BATCH_SIZE;
+            for (int iBatch = 0; iBatch < batches; ++iBatch) {
+                List<DBObject> batch = generateBatch(BATCH_SIZE);
+                WriteResult writeResult = collection.insert(batch);
+
+                if (writeResult.getError() != null) {
+                    log.error(writeResult.getError());
                 }
-                collection.save(document);
+
+                log.info("Count = {}", collection.count());
             }
+
             log.info("complete.");
 
         } catch (MongoException e) {
@@ -46,6 +59,16 @@ public class InsertAHugeNumberOfDocuments implements Command {
         } finally {
             tryToClose(db.getMongo());
         }
+    }
+
+    private List<DBObject> generateBatch(int count) {
+        List<DBObject> batch = new ArrayList<DBObject>();
+        for (int i = 0; i < count; ++i) {
+            BasicDBObject document = new BasicDBObject();
+            document.putAll(nextContact());
+            batch.add(document);
+        }
+        return batch;
     }
